refactor: split useTraitMutation into separate file with error handling
- Move useTraitMutation to its own file per convention (one hook per file)
- Add user-facing error handling with Alert on mutation failure
- Remove duplicate staleTime from useProspectQuery (uses global config)
- Fix variable shadowing: rename 't' to 'trait' in filter callbacks
- Update barrel exports in hooks/index.ts
